requirements.txt
fastapi==0.63.0
gunicorn==20.0.4
uvicorn==0.13.4
uvloop==0.15.2
httptools==0.1.1
asyncpg==0.21.0
h11==0.12.0
psycopg2-binary==2.8.6
starlette==0.13.6
websockets==8.1
aiofiles==0.6.0
adal==1.2.6
azure-common==1.1.26
azure-core==1.12.0
azure-devtools==1.2.0
azure-quantum==0.15.2102.129448
azure-storage-blob==12.8.0
backcall==0.2.0
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
click==7.1.2
colorama==0.4.4
ConfigArgParse==1.3
cryptography==3.4.6
decorator==4.4.2
h11==0.12.0
idna==2.10
ipykernel==5.5.0
ipynb-py-convert==0.4.6
ipython==7.21.0
ipython-genutils==0.2.0
isodate==0.6.0
jedi==0.18.0
jupyter-client==6.1.12
jupyter-core==4.7.1
msal==1.10.0
msrest==0.6.21
msrestazure==0.6.4
multidict==5.1.0
numpy==1.20.1
oauthlib==3.1.0
parso==0.8.1
pickleshare==0.7.5
prompt-toolkit==3.0.17
pycparser==2.20
pydantic==1.8.1
Pygments==2.8.1
PyJWT==2.0.1
python-dateutil==2.8.1
PyYAML==5.4.1
pyzmq==22.0.3
requests==2.25.1
requests-oauthlib==1.3.0
six==1.15.0
starlette==0.13.6
tornado==6.1
traitlets==5.0.5
typing-extensions==3.7.4.3
urllib3==1.26.3
vcrpy==4.1.1
wcwidth==0.2.5
wrapt==1.12.1
yarl==1.6.3